+++ NEWS.md
@@ -15,6 +15,7 @@ Unreleased
 * Introduce `suspenders:db:migrate` task
 * Introduce `suspenders:email` generator
 * Introduce `suspenders:testing` generator
+* Introduce `suspenders:prerequisites` generator
 
 20230113.0 (January, 13, 2023)